Build: Code changes for Emscripten platform support

+++ platform/gucefCORE/src/gucefCORE_CVariantBinarySerializer.cpp
@@ -83,7 +83,7 @@ CVariantBinarySerializer::Serialize( const CVariant& var, UInt32 currentTargetOf
             if ( nrOfBytesWritten != sizeof(payloadSize) )
             {
                 GUCEF_WARNING_LOG( LOGLEVEL_NORMAL, "VariantBinarySerializer:Serialize: payloadSize (" + ToString( payloadSize ) + ") was not written correctly. " + 
-                    ToString( (UInt32) sizeof(payloadSize) ) + " bytes expected, " + ToString( sizeof(nrOfBytesWritten) ) + " bytes written" );
+                    ToString( (UInt32) sizeof(payloadSize) ) + " bytes expected, " + ToString( (UInt32) sizeof(nrOfBytesWritten) ) + " bytes written" );
                 return false;
             }
         }
